split out flag-evaluation and flag-management
